+++ train.py
@@ -18,8 +18,8 @@ def train(epochs=1, vocab_size=391, save=True):
         for i, data in enumerate(pop_rock_train_loader):
             print(f"batch: {i}")
             real_a, real_b = data['bar_a'], data['bar_b']
-            real_a = torch.nn.functional.one_hot(real_a, num_classes=(vocab_size)).float()
-            real_b = torch.nn.functional.one_hot(real_b, num_classes=(vocab_size)).float()
+            real_a = torch.nn.functional.one_hot(real_a, num_classes=(vocab_size))
+            real_b = torch.nn.functional.one_hot(real_b, num_classes=(vocab_size))
             # we may want to feed in as not one_hots and convert to one hots in the model
             real_a, real_b = real_a.to(device), real_b.to(device)
             opt_G_A2B.zero_grad()
